+++ CHANGELOG.md
@@ -9,6 +9,9 @@ CHANGELOG
 - Automatically install missing Python dependencies.
   [#5787](https://github.com/pulumi/pulumi/pull/5787)
 
+- [cli] Ensure `pulumi stack change-secrets-provider` allows rotating the key for a passphrase provider
+  [#5865](https://github.com/pulumi/pulumi/pull/5865/)
+
 ## 2.15.0 (2020-12-02)
 
 - [sdk/python] Add deserialization support for enums.